Makefile: Reduce duplication of rules
Both the all and clean rule targets call the same commands with a
different target.

Create an explicit target for 'modules' and use the '$@' macro from make
to place the target on the commandline.

Update the 'all' target which is the default as it's found first to
depend on the modules target as the default build.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>